Update CNN model init accept channels and mnist arguments
This fix is to update CNN model accepting different values for channels
size and mnist boolean value so that this model can handle both MNIST
and CIFAR10 dataset.

For MNIST dataset, the input size is (1, 28, 28) and we will
additionally pass x thru the self.fc0 to standardise output dimension to
2048 (i.e., to make it compatible with self.fc1).

For CIFAR10 dataset, the self.fc0 layer is not defined and so x wont be
passed thru the self.fc0 layer.